creation de compte et criptage de mot de passe

+++ src/main/java/com/djiguiya/djiguiya/repository/UserRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface UserRepository extends JpaRepository<Utilisateurs,Integer> {
+public interface UserRepository extends JpaRepository<Utilisateurs, Long> {
     Optional<Utilisateurs> findByUsername(String username);
 }
